edited advance used demo docs
